Mini izmene za windows

+++ src/environment.py
@@ -30,9 +30,9 @@ class Ludo(object):
         """
         self.ply = 0
         self.num_players = num_players
-        self.positions = torch.ones(size=(MAX_PLAYERS, TOKENS_PER_PLAYER), dtype=int) * (-1)
-        self.board_state = torch.zeros(BOARD_LENGTH, dtype=int)
-        self.starts = torch.zeros(MAX_PLAYERS, dtype=int)
+        self.positions = torch.ones(size=(MAX_PLAYERS, TOKENS_PER_PLAYER), dtype=torch.long) * (-1)
+        self.board_state = torch.zeros(BOARD_LENGTH, dtype=torch.long)
+        self.starts = torch.zeros(MAX_PLAYERS, dtype=torch.long)
         if self.num_players == 2:
             self.starts[0] = 0
             self.starts[1] = 2 * START_DISTANCE
@@ -45,8 +45,8 @@ class Ludo(object):
             self.starts[1] = 2 * START_DISTANCE
             self.starts[2] = 1 * START_DISTANCE
             self.starts[3] = 3 * START_DISTANCE
-        self.passed = torch.zeros(size=(MAX_PLAYERS, TOKENS_PER_PLAYER), dtype=int)
-        self.home_state = torch.zeros(size=(MAX_PLAYERS, TOKENS_PER_PLAYER), dtype=int)
+        self.passed = torch.zeros(size=(MAX_PLAYERS, TOKENS_PER_PLAYER), dtype=torch.long)
+        self.home_state = torch.zeros(size=(MAX_PLAYERS, TOKENS_PER_PLAYER), dtype=torch.long)
         self.current_player = 0
         self.roll_dice()
 
@@ -130,7 +130,7 @@ class Ludo(object):
         Returns full current board state as a tensor
         :return: Tensor to be passes as input to nn
         """
-        dice_hot = torch.zeros(size=[DICE_MAX])
+        dice_hot = torch.zeros(size=[DICE_MAX], dtype=torch.long)
         dice_hot[self.roll] = 1
         return torch.cat((self.board_state,
                           torch.flatten(self.home_state),
